requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==41.0.0
Django==4.2
django-cors-headers==4.3.0
djangorestframework==3.14.0
idna==3.10
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2025.2
requests==2.31.0
sqlparse==0.5.3
urllib3==2.4.0
